+++ src/BugSnag/Middleware/BugSnagUser.js
@@ -1,16 +1,23 @@
 'use strict'
 
 class BugSnagUser {
-  constructor (Config) {
-    this.appSecret = Config.get('app.appKey')
+  constructor (bugsnag) {
+    this.notifierWrapper = bugsnag
   }
 
-  async handle ({ request, auth }, next) {
-    let _user = await (typeof (auth.getUser) === 'function' ? auth.getUser() : Promise.resolve(null))
+  async handle ({ request, session, auth }, next) {
+    try {
+      let _user = await (typeof (auth.getUser) === 'function' ? auth.getUser() : Promise.resolve(null))
 
-    request.user = _user
+      this.notifierWrapper.setAuthUser(_user)
+      this.notifierWrapper.setContext(request, session)
+    } catch (err) {
+      ;
+    }
 
     await next()
+
+    // this.notifierWrapper.leaveBreadcrumb('"name"', '{metaData}', '"type"', 'timestamp')
   }
 }